feat(jest): 90% coverage for FilSystem.ts
